Handle long path for perl shebang
HostBoot runs perl scripts using shebang, it points to /usr/bin/perl,
but we need a native perl. The solution is in changing shebang inside
HostBoot's perl scripts, but the path to a native perl interpreter may
be too long to fit in shebang format (more than 128 symbols for
kernel 4.x).
So, we create a short link to a native perl and use it in the scripts.

Signed-off-by: Artem Senichev <a.senichev@yadro.com>